+++ objc2/CHANGELOG.md
@@ -23,12 +23,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
       msg_send_id![NSObject::alloc(), init]
   };
   ```
+* Add `Class::class_method`.
 
 ### Changed
 * Allow other types than `&Class` as the receiver in `msg_send_id!` methods
   of the `new` family.
 * **BREAKING**: Changed the `Allocated` struct to be used as `Allocated<T>`
   instead of `Id<Allocated<T>, O>`.
+* Verify the message signature of overriden methods when declaring classes if
+  the `verify_message` feature is enabled.
+* Verify in `declare_class!` that protocols are implemented correctly.
 
 
 ## 0.3.0-beta.3 - 2022-09-01